Extract `taker` and `maker` crates
To actually achieve this, we need to extract another crate
`daemon-shared`. This is not ideal but in the long run, we can try to
get rid of some of this code in there through various means. Some of the
options are documented in the README of the crate.
