+++ daemon/src/lib.rs
@@ -32,6 +32,8 @@ use xtra::message_channel::StrongMessageChannel;
 use xtra::Actor;
 use xtra::Address;
 
+pub use bdk;
+
 pub mod sqlx_ext; // Must come first because it is a macro.
 
 pub mod address_map;
@@ -46,7 +48,6 @@ pub mod connection;
 pub mod db;
 pub mod fan_out;
 pub mod keypair;
-pub mod logger;
 pub mod maker_cfd;
 pub mod maker_inc_connections;
 pub mod model;
@@ -66,7 +67,6 @@ pub mod setup_maker;
 pub mod setup_taker;
 pub mod supervisor;
 pub mod taker_cfd;
-pub mod to_sse_event;
 pub mod tokio_ext;
 mod transaction_ext;
 pub mod try_continue;
