+++ shared-bin/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "shared-bin"
+version = "0.1.0"
+edition = "2021"
+description = "Code that is shared between the daemons but application specific and thus does not go into the `daemon` library."
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+atty = "0.2"
+daemon = { path = "../daemon" }
+rocket = { version = "0.5.0-rc.1", features = ["json"] }
+serde = { version = "1", features = ["derive"] }
+time = "0.3"
+tracing = { version = "0.1" }
+tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt", "ansi", "env-filter", "local-time", "tracing-log", "json"] }
+
+[dev-dependencies]
+serde_test = "1"
