+++ shared-bin/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod logger;
+mod to_sse_event;
+
+pub use crate::to_sse_event::*;
