+++ taker/build.rs
@@ -1,0 +1,14 @@
+use anyhow::Result;
+use vergen::vergen;
+use vergen::Config;
+use vergen::SemverKind;
+
+fn main() -> Result<()> {
+    std::fs::create_dir_all("../maker-frontend/dist/maker")?;
+    std::fs::create_dir_all("../taker-frontend/dist/taker")?;
+
+    let mut config = Config::default();
+    *config.git_mut().semver_kind_mut() = SemverKind::Lightweight;
+
+    vergen(config)
+}
